Added pitting
Added factors for three different levels of corrosion to simulate simple
pitting corrosion of the aluminum foil.  This does not explain the
results totally.  A model that approximates and initiation phase that is
dependent on current but perhaps is an x^2 relationship or something
like that at the beginning and then one with two competing reactions
resulting in a linear phase in the middle would be more accurate.  The
breakdown phase is too random, perhaps, to include.